// gameLogd/src/app/models/home-appliance.model.ts
export interface HomeAppliance {
    id: string;
    title: string;
    description: string;
    brand: string;
    releaseDate: string;
    manufacturer: string;
    distributor: string;
    sales: number;
    rating: number;
    imageUrl: string;
    totalRatingScore: number;
    numRatings: number;
    category: string;
    price: number;
    dimensions: string;
    weight: string;
    powerConsumption: string;
    warranty: string;
    features: string[];
    color: string;
    modelNumber: string;
} 